requirements.txt
beautifulsoup4
chainlit
chromadb
duckduckgo_search==6.3.4
fastapi
faiss-cpu
# flashrank[listwise]
google-generativeai
# google-play-scraper
IPython
ipykernel
langchain==0.3.7
langchain-community==0.3.7
langchain_core
langchain-docling
langchain_google_genai==2.0.4
# langchain_huggingface
langgraph==0.2.48
langserve
langsmith
pandas>=1.3.6
pypdf
PyPDF2
python-dotenv
regex
streamlit
# streamlit-feedback
unstructured[pdf]==0.7.12
# sse_starlette
# uvicorn
-e .
